Rename the abstract_thread module to abstract_generator.
