+++ scripts/abstract_generator.py
